refactor: RoomDao가 Interface에 의존하도록 변경
